Move log entities into special module
